more of the same

+++ php/database/migrations/2019_11_19_211853_create_shipcrewposts_table.php
@@ -16,7 +16,7 @@ class CreateShipCrewPostsTable extends Migration
         Schema::create('ship_crew_posts', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->boolean('is_active');
-            $table->integer('creator_id')->unsigned();
+            $table->bigInteger('creator_id')->unsigned();
             $table->string('invite_code', 6)->unique();
             $table->string('description', 255);
             $table->string('ship', 100);
